endpoint to send ids and accountInfo for specific id account

+++ src/main/java/com/example/DG/bank/system/dto/AccountRequestDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class AccountRequestDTO {
     private long id;
-    private long balance;
+    private Double balance;
     private Type type;
     private long customerid;
     private String customerFirstName;
